requirements.txt
flax
jax==0.4.8
jaxlib==0.4.7
matplotlib==3.7.1
numpy==1.24.2
objax==1.6.0
optax==0.1.4
orbax==0.1.7
pandas==2.0.0
plotly==5.14.1
PyHamcrest==2.0.4
PyQt6==6.5.0
PyYAML==6.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
torch==2.0.0
tqdm==4.65.0
Werkzeug==2.2.3
